requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
ofxparse>=0.21
rapidfuzz>=3.10.0
python-multipart>=0.0.18
httpx>=0.28.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
